moved loading group keys method to shacl test class
